lottery/requirements.txt
django==3.2
djangorestframework
django-cors-headers
django-background-tasks
pillow
mysqlclient